Adding script and button to show and hide password also changing redirecting and adding an alert.
in the script line 10 - 25 I add a js script to show and hide password when the button is click. In line 33 -34 I add the button that when click it will show password and hide password. In 59 - 61 I change the redirecting to a javascript that before redirecting it will alert first the redirect to the new page.